feat: 저장된 역을 삭제하는 기능 구현

+++ src/main/java/subway/service/StationService.java
@@ -30,6 +30,9 @@ public class StationService implements SubwayService {
 
     @Override
     public void delete(String station) {
+        if (stationRepository.findByName(station).isEmpty()) {
+            throw CustomException.from(ErrorMessage.STATION_NOT_FOUND_ERROR);
+        }
         stationRepository.deleteStation(station);
     }
 }
